feat: move collectiontype to state
